+++ test/Grpc.IntegrationTests/OrchestrationErrorHandling.cs
@@ -4,6 +4,7 @@
 using Microsoft.DurableTask.Client;
 using Microsoft.DurableTask.Worker;
 using Xunit.Abstractions;
+using Xunit.Sdk;
 
 namespace Microsoft.DurableTask.Grpc.Tests;
 
@@ -11,11 +12,8 @@ namespace Microsoft.DurableTask.Grpc.Tests;
 /// Integration tests that are designed to exercise the error handling and retry functionality
 /// of the Durable Task SDK.
 /// </summary>
-public class OrchestrationErrorHandling : IntegrationTestBase
+public class OrchestrationErrorHandling(ITestOutputHelper output, GrpcSidecarFixture sidecarFixture) : IntegrationTestBase(output, sidecarFixture)
 {
-    public OrchestrationErrorHandling(ITestOutputHelper output, GrpcSidecarFixture sidecarFixture)
-        : base(output, sidecarFixture)
-    { }
 
     /// <summary>
     /// Tests the behavior and output of an unhandled exception that originates from an activity.
@@ -71,8 +69,11 @@ public class OrchestrationErrorHandling : IntegrationTestBase
     /// This is different from <see cref="UnhandledActivityException"/> in that the source of the
     /// exception is in the orchestrator code directly, and not from an unhandled activity task.
     /// </remarks>
-    [Fact]
-    public async Task UnhandledOrchestratorException()
+    [Theory]
+    [InlineData(typeof(ApplicationException))] // built-in exception type
+    [InlineData(typeof(CustomException))] // custom exception type
+    [InlineData(typeof(XunitException))] // 3rd party exception type
+    public async Task UnhandledOrchestratorException(Type exceptionType)
     {
         string errorMessage = "Kah-BOOOOOM!!!"; // Use an obviously fake error message to avoid confusion when debugging
         string? expectedCallStack = null;
@@ -85,7 +86,7 @@ public class OrchestrationErrorHandling : IntegrationTestBase
                 // The Environment.StackTrace and throw statements need to be on the same line
                 // to keep line numbers consistent between the expected stack trace and the actual stack trace.
                 // Also need to remove the top frame from Environment.StackTrace.
-                expectedCallStack = Environment.StackTrace.Replace("at System.Environment.get_StackTrace()", string.Empty).TrimStart(); throw new Exception(errorMessage);
+                expectedCallStack = Environment.StackTrace.Replace("at System.Environment.get_StackTrace()", string.Empty).TrimStart(); throw MakeException(exceptionType, errorMessage)!;
             }));
         });
 
@@ -99,11 +100,12 @@ public class OrchestrationErrorHandling : IntegrationTestBase
 
         Assert.NotNull(metadata.FailureDetails);
         TaskFailureDetails failureDetails = metadata.FailureDetails!;
-        Assert.Equal(typeof(Exception).FullName, failureDetails.ErrorType);
+        Assert.Equal(exceptionType.FullName, failureDetails.ErrorType);
         Assert.Equal(errorMessage, failureDetails.ErrorMessage);
         Assert.NotNull(failureDetails.StackTrace);
         Assert.NotNull(expectedCallStack);
         Assert.Contains(expectedCallStack![..300], failureDetails.StackTrace);
+        Assert.True(failureDetails.IsCausedBy(exceptionType));
     }
 
     /// <summary>
@@ -149,10 +151,10 @@ public class OrchestrationErrorHandling : IntegrationTestBase
     }
 
     [Theory]
-    [InlineData(1)]
-    [InlineData(2)]
-    [InlineData(10)]
-    public async Task RetryActivityFailuresCustomLogic(int expectedNumberOfAttempts)
+    [InlineData(1, typeof(ApplicationException))] // 1 attempt, built-in exception type
+    [InlineData(2, typeof(CustomException))] // 2 attempts, custom exception type
+    [InlineData(10, typeof(XunitException))] // 10 attempts, 3rd party exception type
+    public async Task RetryActivityFailuresCustomLogic(int expectedNumberOfAttempts, Type exceptionType)
     {
         string errorMessage = "Kah-BOOOOOM!!!"; // Use an obviously fake error message to avoid confusion when debugging
 
@@ -171,8 +173,8 @@ public class OrchestrationErrorHandling : IntegrationTestBase
                 return false;
             }
 
-            // This handler only works with ApplicationException
-            if (!retryContext.LastFailure.IsCausedBy<ApplicationException>())
+            // This handler only works with CustomException
+            if (!retryContext.LastFailure.IsCausedBy(exceptionType))
             {
                 return false;
             }
@@ -194,7 +196,7 @@ public class OrchestrationErrorHandling : IntegrationTestBase
                 .AddActivityFunc("Foo", (TaskActivityContext context) =>
                 {
                     actualNumberOfAttempts++;
-                    throw new ApplicationException(errorMessage);
+                    throw MakeException(exceptionType, errorMessage);
                 }));
         });
 
@@ -213,10 +215,10 @@ public class OrchestrationErrorHandling : IntegrationTestBase
     /// Tests retry policies for sub-orchestration calls.
     /// </summary>
     [Theory]
-    [InlineData(1)]
-    [InlineData(2)]
-    [InlineData(10)]
-    public async Task RetrySubOrchestrationFailures(int expectedNumberOfAttempts)
+    [InlineData(1, typeof(ApplicationException))] // 1 attempt, built-in exception type
+    [InlineData(2, typeof(CustomException))] // 2 attempts, custom exception type
+    [InlineData(10, typeof(XunitException))] // 10 attempts, 3rd party exception type
+    public async Task RetrySubOrchestrationFailures(int expectedNumberOfAttempts, Type exceptionType)
     {
         string errorMessage = "Kah-BOOOOOM!!!"; // Use an obviously fake error message to avoid confusion when debugging
 
@@ -237,7 +239,7 @@ public class OrchestrationErrorHandling : IntegrationTestBase
                 .AddOrchestratorFunc("BustedSubOrchestrator", context =>
                 {
                     actualNumberOfAttempts++;
-                    throw new ApplicationException(errorMessage);
+                    throw MakeException(exceptionType, errorMessage);
                 }));
         });
 
@@ -249,13 +251,18 @@ public class OrchestrationErrorHandling : IntegrationTestBase
         Assert.Equal(instanceId, metadata.InstanceId);
         Assert.Equal(OrchestrationRuntimeStatus.Failed, metadata.RuntimeStatus);
         Assert.Equal(expectedNumberOfAttempts, actualNumberOfAttempts);
+        Assert.NotNull(metadata.FailureDetails);
+        Assert.Contains(errorMessage, metadata.FailureDetails!.ErrorMessage);
+
+        // The root orchestration failed due to a failure with the sub-orchestration, resulting in a TaskFailedException
+        Assert.True(metadata.FailureDetails.IsCausedBy<TaskFailedException>());
     }
 
     [Theory]
-    [InlineData(1)]
-    [InlineData(2)]
-    [InlineData(10)]
-    public async Task RetrySubOrchestratorFailuresCustomLogic(int expectedNumberOfAttempts)
+    [InlineData(1, typeof(ApplicationException))] // 1 attempt, built-in exception type
+    [InlineData(2, typeof(CustomException))] // 2 attempts, custom exception type
+    [InlineData(10, typeof(XunitException))] // 10 attempts, 3rd party exception type
+    public async Task RetrySubOrchestratorFailuresCustomLogic(int expectedNumberOfAttempts, Type exceptionType)
     {
         string errorMessage = "Kah-BOOOOOM!!!"; // Use an obviously fake error message to avoid confusion when debugging
 
@@ -274,8 +281,8 @@ public class OrchestrationErrorHandling : IntegrationTestBase
                 return false;
             }
 
-            // This handler only works with ApplicationException
-            if (!retryContext.LastFailure.IsCausedBy<ApplicationException>())
+            // This handler only works with CustomException
+            if (!retryContext.LastFailure.IsCausedBy(exceptionType))
             {
                 return false;
             }
@@ -297,7 +304,7 @@ public class OrchestrationErrorHandling : IntegrationTestBase
                 .AddOrchestratorFunc("BustedSubOrchestrator", context =>
                 {
                     actualNumberOfAttempts++;
-                    throw new ApplicationException(errorMessage);
+                    throw MakeException(exceptionType, errorMessage);
                 }));
         });
 
@@ -310,6 +317,10 @@ public class OrchestrationErrorHandling : IntegrationTestBase
         Assert.Equal(OrchestrationRuntimeStatus.Failed, metadata.RuntimeStatus);
         Assert.Equal(expectedNumberOfAttempts, retryHandlerCalls);
         Assert.Equal(expectedNumberOfAttempts, actualNumberOfAttempts);
+
+        // The root orchestration failed due to a failure with the sub-orchestration, resulting in a TaskFailedException
+        Assert.NotNull(metadata.FailureDetails);
+        Assert.True(metadata.FailureDetails!.IsCausedBy<TaskFailedException>());
     }
 
     [Theory]
@@ -351,5 +362,15 @@ public class OrchestrationErrorHandling : IntegrationTestBase
 
         // The retry handler should never get called for a missing activity or sub-orchestrator exception
         Assert.Equal(0, retryHandlerCalls);
+    }
+
+    static Exception MakeException(Type exceptionType, string message)
+    {
+        // We assume the contructor of the exception type takes a single string argument
+        return (Exception)Activator.CreateInstance(exceptionType, message)!;
+    }
+
+    class CustomException(string message) : Exception(message)
+    {
     }
 }